chat_service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35

# Postgres client for mirroring Smart Access into RAG database
pg8000==1.30.5
